requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gradio==4.39.0
gradio_client==1.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pinecone==4.0.0
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
ruff==0.5.4
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
tqdm==4.66.4
transformers==4.43.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
